updated react-navigation to v4.0
