+++ package.json
@@ -30,9 +30,13 @@
     "react-native": "https://github.com/expo/react-native/archive/sdk-34.0.0.tar.gz",
     "react-native-elements": "^1.2.5",
     "react-native-gesture-handler": "~1.3.0",
+    "react-native-reanimated": "~1.1.0",
     "react-native-touchable-scale": "^2.0.0",
     "react-native-web": "^0.11.4",
-    "react-navigation": "^3.11.1"
+    "react-navigation": "^4.0.10",
+    "react-navigation-drawer": "^1.4.0",
+    "react-navigation-stack": "^1.7.3",
+    "react-navigation-tabs": "^1.2.0"
   },
   "devDependencies": {
     "@expo/webpack-config": "^0.7.3",
