+++ src/tabs/lists.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import { createStackNavigator } from 'react-navigation';
+import { createStackNavigator } from 'react-navigation-stack';
 import { Icon } from 'react-native-elements';
 
 import ListsHome from '../views/lists_home';
